New glyphs update
Note:
In the first stage, only the glyph that is the commonly used character will be implemented. (The definition of commonly used character is the character listed in cm-chara-list)
For Simplified Chinese specific characters (Not include the non-orthodox traditional character, including but not limited to 国, 会, 区etc. ): Partially implement in the first release.
For new glyph of CL Variant: The glyphs form standard is following the Japanese Kyuujitai standard, which is considered as the de facto standard of the glyph forms that were used in the past.
For new glyph of CM Variant: Minor different of glyph shape such as 月 may not be implement.